Former-commit-id: 00bb53830f25ea050741ab07200676593b706870

+++ WhirlyGlobeSrc/WhirlyGlobeComponentTester/WhirlyGlobeComponentTester/ConfigViewController.m
@@ -183,7 +183,7 @@
     return cs.sectionName;
 }
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     if (section >= [_values count])
         return 0;
